the test of whether ffinit/create_file would overwrite an existing file was not win32 safe
The test created a FITS handle to a new file, then attempted to
delete/create the file while the first FITS handle (and thus the
underlying file handle )was still active.

This is legal on *nix, but not on Windows:

   https://learn.microsoft.com/en-us/windows/win32/fileio/closing-and-deleting-files

   A file cannot be deleted until all handles to it are closed. If a file cannot be deleted, its name cannot be reused

The tests are reordered so that there are no open filehandles when the test for creating an existing file is done.